Configure the SOAP client to always return an array for a sequence
PHPs SOAP client has inconsistent default behaviour. If you receive a single element sequence, the client will set the property to the sequence member. If you have multiple sequence elements the property is an array of members instead. This is why Utils::EnsureIsArray() is used in the AvaTax code to 'fix' the problem. The better fix is to use the feature flag that turns on the correct behaviour of always returning an array for a sequence.